Extract calendar from SplitView
